quick-verification/snippets/requirements-rag.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
qdrant-client>=1.7.0
numpy>=1.24.0
flask>=3.0.0
